Remove BRMO Transform action from CVnHR
