+++ QNH.Overheid.KernRegister.Beheer/Controllers/SearchController.cs
@@ -152,7 +152,6 @@ namespace QNH.Overheid.KernRegister.Beheer.Controllers
 
                         var brmoSyncService = nestedContainer.GetInstance<IBrmoSyncService>();
                         brmostatus = brmoSyncService.UploadXDocumentToBrmo(xDoc);
-                        brmoSyncService.Transform(kvkNummer);
                     }
                     catch (Exception ex)
                     {
